feat(startup) : add version info and header to the top of startup log file

+++ src/FlexSearch.Core/Bootstrapper.fs
@@ -91,7 +91,7 @@ module FactoryService =
             
             member __.GetMetaData(moduleName) = 
                 match getModuleByName (moduleName, true) with
-                | Choice2Of3(x) -> ok(x)
+                | Choice2Of3(x) -> ok (x)
                 | _ -> fail <| ModuleNotFound(moduleName, moduleTypeName)
             
             member __.ModuleExists(moduleName) = 
@@ -110,8 +110,7 @@ module FactoryService =
                             let pluginValue = plugin.Value.Value
                             modules.Add(pluginName, pluginValue)
                             moduleNames.Add(pluginName)
-                        with e -> 
-                            Logger.Log <| PluginLoadFailure(pluginName, typeof<'T>.FullName, exceptionPrinter e)
+                        with e -> Logger.Log <| PluginLoadFailure(pluginName, typeof<'T>.FullName, exceptionPrinter e)
                 Logger.Log <| PluginsLoaded(typeof<'T>.FullName, moduleNames)
                 modules
     
@@ -158,7 +157,7 @@ module Main =
         builder |> FactoryService.registerSingleInstance<DemoIndexService, DemoIndexService>
         builder |> FactoryService.registerSingleInstance<EventAggregrator, EventAggregrator>
         builder.Build()
-    
+
 //            let indexService = container.Resolve<IIndexService>()
 // Close all open indices
 //            match indexService.GetAllIndex() with
@@ -166,15 +165,14 @@ module Main =
 //                for registeration in regs do
 //                    indexService.CloseIndex(registeration.IndexName) |> ignore
 //            | _ -> ()
-
-module Interop =
+module Interop = 
     open System
     open System.Linq
     open System.Runtime.InteropServices
     open System.Reflection
     open System.Threading
-
-    type HTTP_SERVER_PROPERTY =
+    
+    type HTTP_SERVER_PROPERTY = 
         | HttpServerAuthenticationProperty = 0
         | HttpServerLoggingProperty = 1
         | HttpServerQosProperty = 2
@@ -187,30 +185,25 @@ module Interop =
         | HttpServerListenEndpointProperty = 9
         | HttpServerChannelBindProperty = 10
         | HttpServerProtectionLevelProperty = 11
-
+    
     [<DllImport("httpapi.dll", CallingConvention = CallingConvention.StdCall)>]
-    extern uint32 HttpSetRequestQueueProperty(
-            CriticalHandle requestQueueHandle,
-            HTTP_SERVER_PROPERTY serverProperty,
-            uint32& pPropertyInfo,
-            uint32 propertyInfoLength,
-            uint32 reserved,
-            IntPtr pReserved)
-
+    extern uint32 HttpSetRequestQueueProperty(CriticalHandle requestQueueHandle, HTTP_SERVER_PROPERTY serverProperty, uint32& pPropertyInfo, uint32 propertyInfoLength, uint32 reserved, IntPtr pReserved)
+    
     // Adapted from:
     // http://stackoverflow.com/questions/15417062/changing-http-sys-kernel-queue-limit-when-using-net-httplistener
     /// Sets the request queue length of a HTTP listener
-    let setRequestQueueLength(listener: System.Net.HttpListener, len: uint32) =
+    let setRequestQueueLength (listener : System.Net.HttpListener, len : uint32) = 
         let listenerType = typeof<System.Net.HttpListener>
         let mutable length = len
         let requestQueueHandleProperty = 
-            listenerType.GetProperties(BindingFlags.NonPublic ||| BindingFlags.Instance).First(fun p -> p.Name = "RequestQueueHandle")
-
+            listenerType.GetProperties(BindingFlags.NonPublic ||| BindingFlags.Instance)
+                        .First(fun p -> p.Name = "RequestQueueHandle")
         let requestQueueHandle = requestQueueHandleProperty.GetValue(listener) :?> CriticalHandle
-        let result = HttpSetRequestQueueProperty(requestQueueHandle, HTTP_SERVER_PROPERTY.HttpServerQueueLengthProperty, &length, Marshal.SizeOf(len) |> uint32, 0u, IntPtr.Zero);
-
-        if result <> 0u then
-            failwithf ""
+        let result = 
+            HttpSetRequestQueueProperty
+                (requestQueueHandle, HTTP_SERVER_PROPERTY.HttpServerQueueLengthProperty, &length, 
+                 Marshal.SizeOf(len) |> uint32, 0u, IntPtr.Zero)
+        if result <> 0u then failwithf ""
 
 /// Used by windows service (top shelf) to start and stop windows service.
 [<Sealed>]
@@ -229,79 +222,97 @@ type NodeService(serverSettings : Settings.T, testServer : bool) =
             httpServer <- new OwinServer(generateRoutingTable handlerModules, port)
             httpServer.Start()
         with e -> printfn "%A" e
-        
+    
     member __.Stop() = httpServer.Stop()
 
-type EventTextFormatter() =
+type EventTextFormatter() = 
     interface Formatters.IEventTextFormatter with
-        member __.WriteEvent(eventEntry, writer) =
+        member __.WriteEvent(eventEntry, writer) = 
             writer.WriteLine(sprintf "[%s] %s" (eventEntry.Schema.Level.ToString()) eventEntry.FormattedMessage)
 
-module StartUp =
+module StartUp = 
+    open System.Reflection
     
-    let (!>) (msg: string) = Logger.Log(msg, MessageKeyword.Startup, MessageLevel.Verbose)
+    let version = Assembly.GetExecutingAssembly().GetName().Version.ToString()
+    
+    /// Header text to be displayed at the beginning of the program
+    let headerText = 
+        let text = """FlexSearch Server
+Flexible and fast search engine for the .Net Platform
+------------------------------------------------------------------
+Version : {version}
+Copyright (C) 2010 - {year} - FlexSearch
+------------------------------------------------------------------
 
+"""
+        text.Replace("{version}", version).Replace("{year}", DateTime.Now.Year.ToString())
+    
+    let (!>) (msg : string) = Logger.Log(msg, MessageKeyword.Startup, MessageLevel.Verbose)
+    
     /// Checks if the application is in interactive user mode?
-    let isInteractive =
+    let isInteractive = 
         let args = Environment.GetCommandLineArgs()
         Environment.UserInteractive || (notNull args && args.Length > 0)
-
+    
     let private consoleSink = ConsoleLog.CreateListener(new EventTextFormatter())
+    
     let private rollingFileSink = 
-        let fileName = Constants.LogsFolder +/ "startup-log.txt" 
-        if File.Exists(fileName) then
-            File.Delete(fileName)
+        let fileName = Constants.LogsFolder +/ "startup-log.txt"
+        if File.Exists(fileName) then File.Delete(fileName)
+        File.AppendAllText(fileName, headerText)
         FlatFileLog.CreateListener(fileName, new EventTextFormatter())
-
+    
     /// Initialize the listeners to be used across the application
-    let initializeListeners() =
-        if isInteractive then
+    let initializeListeners() = 
+        if isInteractive then 
             // Only use console listener in user interactive mode
             consoleSink.EnableEvents(LogService.GetLogger(), EventLevel.LogAlways)
-
         // Write all start up events to a specific file. This is helpful in case ETW is not
         // setup properly. The slight overhead of writing to two sinks is negligible.
-        rollingFileSink.EnableEvents(LogService.GetLogger(), EventLevel.LogAlways, LogService.Keywords.Startup ||| LogService.Keywords.Node)
-            
+        rollingFileSink.EnableEvents
+            (LogService.GetLogger(), EventLevel.LogAlways, LogService.Keywords.Startup ||| LogService.Keywords.Node)
+    
     /// To improve CPU utilization, increase the number of threads that the .NET thread pool expands by when
     /// a burst of requests come in. We could do this by editing machine.config/system.web/processModel/minWorkerThreads,
     /// but that seems too global a change, so we do it in code for just our AppPool. More info: 
     /// http://support.microsoft.com/kb/821268
     /// http://blogs.msdn.com/b/tmarq/archive/2007/07/21/asp-net-thread-usage-on-iis-7-0-and-6-0.aspx
     /// http://blogs.msdn.com/b/perfworld/archive/2010/01/13/how-can-i-improve-the-performance-of-asp-net-by-adjusting-the-clr-thread-throttling-properties.aspx
-    let maximizeThreads() =
+    let maximizeThreads() = 
         let newMinWorkerThreads = 10
         let (minWorkerThreads, minCompletionPortThreads) = ThreadPool.GetMinThreads()
         ThreadPool.SetMinThreads(Environment.ProcessorCount * newMinWorkerThreads, minCompletionPortThreads)
     
     /// Capture all un-handled exceptions
-    let subscribeToUnhandledExceptions() =
+    let subscribeToUnhandledExceptions() = 
         AppDomain.CurrentDomain.UnhandledException.Subscribe(fun x -> 
             Logger.Log(sprintf "%A" x.ExceptionObject, MessageKeyword.Node, MessageLevel.Critical)
-            if isInteractive then
-                printfn "The application has encountered a critical error and will shutdown. Please refer to the Startup-Log.txt under logs folder to get more information."
+            if isInteractive then 
+                printfn 
+                    "The application has encountered a critical error and will shutdown. Please refer to the Startup-Log.txt under logs folder to get more information."
                 printfn "Press any key to continue . . ."
-                Console.ReadKey() |> ignore
-        ) |> ignore
+                Console.ReadKey() |> ignore)
+        |> ignore
     
     /// Load server settings
-    let loadSettings() =
-        match Settings.create(Constants.ConfFolder +/ "Config.ini") with
+    let loadSettings() = 
+        match Settings.create (Constants.ConfFolder +/ "Config.ini") with
         | Ok(s) -> new Settings.T(s)
         | Fail(e) -> 
             Logger.Log
-                ("Error parsing 'Config.ini' file.", ValidationException(e), MessageKeyword.Startup, MessageLevel.Critical)
+                ("Error parsing 'Config.ini' file.", ValidationException(e), MessageKeyword.Startup, 
+                 MessageLevel.Critical)
             failwithf "%s" (e.ToString())
     
     /// Load all plug-ins from the plug-ins folder
-    let loadAllPlugins() =
+    let loadAllPlugins() = 
         for file in Directory.EnumerateFiles(Constants.PluginFolder, "*.dll", SearchOption.TopDirectoryOnly) do
             try 
                 System.Reflection.Assembly.LoadFile(file) |> ignore
             with e -> Logger.Log("Error loading plug-in library.", e, MessageKeyword.Startup, MessageLevel.Warning)
     
     /// Load all the server components and return settings
-    let load() =
+    let load() = 
         // NOTE: The order of operations below is very important 
         // It is important to initialize Listeners first otherwise logging services will not be available 
         initializeListeners()
